+++ apps/web/components/common/Sidebar/index.tsx
@@ -21,7 +21,7 @@ export default function Sidebar(props: SidebarProps): JSX.Element | null {
   return createPortal(
     <>
       {/* 배경 오버레이 */}
-      {isOpen && <div className="fixed inset-0 z-40" onClick={onClose} />}
+      {isOpen ? <div className="fixed inset-0 z-40" onClick={onClose} /> : null}
 
       <div
         className={`min-w-400 border-l-1 border-custom-black/20 fixed right-0 top-0 z-50 hidden h-full transform bg-white shadow-lg transition-transform duration-300 md:block ${
